Add doxygen infrastructure
Adds support for documentation generated by Doxygen as html including a
custom theme. The theme is based on a Bootstrap with the help of
https://github.com/Velron/doxygen-bootstrapped and
https://github.com/StratifyLabs/Doxygen-Bootstrap.

No generated output is bundled currently, as it contains binary images
and such that I'd rather avoid having in the repository.

To generate the documentation, simply run `doxygen doc/Doxyfile` from
the source directory.