Add "Blender Flat" style for testing, add style toggle button
As always there are some things here that are not so nice (e.g. read
comment for RadioButton class).
Stage handles the style radio buttons in a pretty hackisch way, but I
think this is okay since we need to make things work for Blender
specifically.

The added "Blender Flat" style is meant for testing purposes, it's quite
similar to "Blender Classic".