Add an initial text box button
Text editing and inputting isn't possible yet, have to figure out the
best way to do this still. On the long run it might be reasonable to
have a text-editing API within bWidgets that can be used outside of the
library too. However, this can be done later, for the Blender
integration we can just call Blender's text editing code.
To exit the pseudo text editing mode of the button, just right-click it.