bWidgetsDemo: Add & use simple layout system
This commit adds a layout system loosely based on Blender's. It has a
small number of features:
* Automatic row layout (items added left to right)
* Automatic column layout (items added top to bottom)
* Layout nesting (e.g. columns within a row)
* Alignment support (stitching widgets together like in Blender)

Of course there are many things that could be improved, but I'm quite
happy with it for now. It's only for testing purposes anyway and for
that it's more than fine.

Also changes the default UI of the demo application to show the
style-selector, some push buttons (excerpt from Blender 3D View
toolshelf) and a text-edit button. These buttons are managed by the
layout system.