Use functors for generic applying of widget edits
Using functors, we can completely avoid storing any data the widget
represents within the wigdet itself. This data may vary a lot in type
and be application specific, so it makes sense to abstract it away.

Now, widgets don't know anything about the data they represent. Instead,
they register functors for storing references to and managing the data.

As an example, Blender would register functors for applying RNA property
changes which would store the pointer/property combination. When the
widget value changes, the functor would execute and update the property
with the new value.

Qt does this with their signal-slot system, they don't use functors for
it. In future we may switch to that, but functors should work just fine
for us.