Manage drawing using a paint-engine (bwPaintEngine)
This is quite a bit nicer than using callbacks for drawing. It allows
having multiple drawing backends and swapping them (on runtime even),
plus, it's a nicer interface to work with.
Later we could use it to do things like drawing to image buffers or
textures (e.g. for docs or thumbnail-previews). Or even more fancy stuff
like exporting to HMTL & CSS (useful for online mockups/prototypes). One
'simply' has to write a paint-engine to do that.

Demo application uses a GawainPaintEngine for drawing.