Change license of Demo application to GPL 2 or later
Reason for this is mainly to avoid further licensing issues when adding
components based on Blender (e.g. Blender code or icons). With the MIT license,
we'd have to ask every contributor of GPL components for their permission to use
their contributions. With the demo app being GPL2+ we can use their
contributions without permission.
Also, while I don't think GPL makes sense for bWidgets itself, it's fine for the
demo app. I support the philosophy behind the GPL, so I like to use it where
reasonable.

It seems generally suggested to make copyright clear on a per file basis - at
least for the original code. Thus I decided to add copyright notices to all
files. Some code is based on Blender code, so I tried to find out the copyright
holders of it and added them to the copyright notice. I can not guarantee that
these copyright notices are correct, but as I understand they are only acting as
a good hint anyway.