Add own implementation of C++17's std::optional (bwOptional)
The implementation of `bwOptional` should follow the C++17 standard of
std::optional as far as possible with C++11. Inconsistencies (other than known
ones as listed in the main `bwOptional` comment block) should be reported.
Includes extensive test coverage.
For information on how to use `bwOptional`, please refer to the existing
documentation of `std::optional`.

Both the implementation and the tests for `bwOptional` are based on Clang's
implementation/tests of `std::optional`. I added the required copyright notice.

Note that the `bwOptional` implementation is quite complicated and not easy to
follow. I tried to keep it simple but there was no way to do so while following
all the requirements the standards sets (for good reasons). Maybe a full-blown
implementation of the standard is a bit overkill, however it was a nice practice
for me to get into std level library development ;)
I also considered using existing implementations/variations, such as
`Boost::Optional`, but I couldn't just include a single header and adding
dependencies is something I'd still like to avoid.

Also: Splits bwUtil.h into multiple files and puts them together into
bwidgets/utils.