Use bwOptional and bwPointer throughout bWidgets and the demo
This moves us more into the direction of modern C++, where ownership is made
clear, memory-leaks are a thing of the past (implicit freeing thanks to RAII),
code is generally more exception safe, semantics of function arguments and
return-values are more clear (no more paranoid NULL-checks or "return -1 in case
of faile" thanks to bwOptional), ...

One thing that annoys me is having to use
`bwOptional<std::reference_wrapper<Foo>>` pretty often. Since passing a
reference is the same as passing a `const *` we could use that instead.