Styling refactor: Register style property types globally
It doesn't make much sense to register property type data per widget, since that
unnecessarily duplicates it. E.g. the "background-color" will always be a color
property that shouldn't inherit by default.
The global property type map isn't used yet. I'm planning more, bigger changes
to improve our style system and to support more CSS features like inheritance.

Now the CSS parser will be able to query information on the property from the
name it parsed, so it knows what value types to accept. Right now the property
type is guessed from the parsed value and we only fail when applying it to the
widget fails. The planned new approach separates concerns better, so it will be
able to fail earlier, be easier to extend and simpler to unit test.

I've also introduced the concept of style-structs so that related properties
that share the same characteristics (i.e. inheritance) can be grouped together.
This can later be used to manage CSS shorthand properties nicely.
Web-browsers use this style-struct approach too, so worth giving it a try.