Demo App: Custom iterator for iterating widgets in a layout (sub-)tree
Iterating widgets in a layout is now as simple as:
```
for (bwWidget& widget : layout) {
    // Do stuff...
}
```

Or if hidden widgets should be included:
```
for (bwWidget& widget : WidgetIterator::withHidden(layout)) {
    // Do stuff...
}
```

Note that this also works for any subtree, e.g. to iterate only the widgets
within a panel:
```
for (bwWidget& widget : panel_layout) {
    // Do stuff...
}
```

This may very well end up in bWidgets itself, so decided to test this well from
the get-go. Gcov reports 100% coverage.

Also did some changes to get the label text from a widget. Changes in the layout
code were also needed.