Docs: Update design overview for new screen-graph/pipeline design
The design overview is now the most important doc as it describes the
very basics of how bWidgets is designed. Especially in regards to the
screen-graph and the pipeline. That said, most of the page is still
TODO.
I also carefully wrote down design rationales, which are hidden by
default but can be opened easily. Had to do some JavaScript tricks to
get that to work.