Introducing: The screen-graph
The screen-graph is going to be the central data-structure of bWidgets.
It is a tree, defining the hierarchy and owning all widgets and layouts.
Yet, its nodes are designed to be small, helper classes give them
functionality. I.e. it can be built conveniently via a screen-graph
`Builder`

An overview on what role the screen-graph plays, and a design rationale
for it can be found in the documentation:
https://julianeisel.github.io/bWidgets/md_docs_bWidgets_design_overview.html

This commit contains:
* Initial screen-graph node classes
* Builder class for screen-graph construction
* Convenient, stack friendly pre-order iterator
* Tests for screen-graph iterator
* New `bwLayoutInterface` to allow storing custom layouts in the screen-graph
* Documentation updates
* Changes to demo application to make it work with screen-graphs