Add Scroll-View widget type and many changes to support them
Introduces a `bwScrollView` widget which should automatically do all handling of scrollbars based on bounds of child widgets. All in all this simplifies and improves things quite a bit, even if there are still some annoyances.
Currently only vertical scrolling supported, horizontal should be easy to add but isn't needed right now.

Main changes in bWidgets itself:
* Adds abstract `bwContainerWidget` class as base for all widgets supporting usage as container.
* Introduces event bubbling, used only for scroll-wheel events for now: If an event is not "swallowed" (`event.swallow()`), it is sent up the screengraph for ancestors to handle it (until swallowed). That way the scroll wheel works even when hovering other widgets. Should be done for other, if not all event types.
* Introduces content masks: Rectangles for container widgets for clipping children.
* Adds a helper class to draw the screen-graph, i.e. `bwScreenGraph::draw()`. It handles logic for content masks.
* Refactors anti-aliased outline drawing, so that non-anti-aliased outlines can be drawn too.
* Adds handling for scroll-wheel events to widgets.
* Clicking inside scroll-bar, but outside of the scroll-value indicator jumps the view, like in Blender (and everywhere else in the world).

Done in the handling branch because a) rewriting scroll logic like this was needed anyway to keep them working with new handling, and b) changes in the branch make sense for this feature.