Demo App: Initial RNA-like property system
Used to figure out and test how bWidgets could interact with a system
like RNA, without knowing of its existence. Also gives a better idea of
how constructing a screen-graph could look like with such a system.

What we have here is an initial working version which can be simplified
and refined.