Update to newer Gawain version & smaller cleanup/fixes
Unfortunately updating to rather new versions of the Blender GPU module
(basically the successor of Gawain) isn't easily possible. There are a
number of calls to Blender functions. The main reason I'm updating
Gawain is that I'm trying hard for a while to fix a crash in the
graphics driver on exit. Unfortunately it's still not gone, but why not
update the lib if I already did all the work for it?