Demo App: Support hiDPI drawing
The DPI scaling is only applied for drawing, not within bWidgets itself. That
seems to work just fine. It's interesting that we may be able to apply hiDPI
scaling only at this very late and localized stage, without having to apply it
all over the place within bWidgets.

Note that icon PNGs are still just 16x16px and upscaled for hiDPI.