Cleanup: Refactor widget type identification
* Removes bwWidget::Type
* Add `widget_cast<>()` variant for references
* Remove unnecessary try-catch in `widget_cast<>()

Using `widget_cast<>()` is now the preferred way to check the type of a widget
(if it's not of the requested type, `nullptr` is returned).
I've explained the reasoning in a comment for the function. In short I think
it's safer (type check performed by compiler) and more extensible (a widget
type can be defined in a single file, no central register needed).
Many `dynamic_cast<>()`'s (which is what `widget_cast<>()` uses) are typically
considered a code smell, because OOP gives "better" ways to customize an
object's behavior. And indeed, the widgets and their relatives (e.g. their
widget specific handlers) should contain most customization needed. So
eventually, there should be very few widget type checks. For now we still need
to do it in a couple of places though.