Builder: Build sub-layouts using callbacks, avoids exposing state
E.g. this:
```
  LayoutNode& root_layout = builder.getActiveLayout()

  builder.addLayout<RowLayout>();
  builder.addWidget<bwCheckbox>()
      .setLabel("Foo");
  builder.addWidget<bwCheckbox>()
      .setLabel("Bar")
      .hide();

  builder.setActiveLayout(root);
```
can be turned into this:
```
  builder.buildLayout<RowLayout>([](Builder& builder) {
      builder.addWidget<bwCheckbox>()
          .setLabel("Foo");
      builder.addWidget<bwCheckbox>()
          .setLabel("Bar")
          .hide();
  });
```

This makes the stucture much more explicit, and constructing a layout becomes
stateless (no need for the calling code to manage the active layout of the
builder). It also generalizes the layout construction in a way that should make
more complex hierarchies simpler to build.