Initial work to reconstruct the screen-graph on redraws
I decided to go for a design that allows an immediate-mode like creation of
layouts, similar like things work in Blender already. E.g.:
```
if (some_condition == true) {
  create_button();
}
```

That is different from the more retained-mode layout creation, where the
individual layout elements have to be permanently stored and you explicitly
hide and unhide them as needed.

For the immediate-mode like layout creation to work, the entire layout has to
be reconstructed on every redraw. bWidgets will make sure the state of
individual widgets is kept over redraws, even if they are reconstructed. It's
important that widgets can be unambiguously identified for this. Other toolkits
do this with identifier names or such, I think we should keep doing what
Blender is doing, which is identifying widgets by comparing the data they
store. This commit works towards these things.

I'm planning to write design documents on this. It would be an important
characteristic of the bWidgets design and impacts the creation of custom widget
types. So it should be documented well.

Note that this isn't ready for use at all, it crashes on the first redraw.