Add trival to use system for persistent screen-graph node references
One issue with recreating the UI on every redraw is that widgets can't be
stored permanently. We have the same issue in Blender already and have some
ways around it, but it's quite a hassle. So I always wanted to look into a
better system for bWidgets.

This implements a super simple to use `PersistentNodeRef` type to store
references that are automatically updated on reconstructions.
Usage is as simple as this:
```
PersistentNodeRef node_ref = ScreenGraph::make_persistent_ref(screen_graph, node, "Identifier for debugging");
```

The node references are added to a registry in the screen-graph to update them.
Copies of the node references are updated as well and the original reference
can be destroyed or moved from. Internally, a `std::shared_ptr<Node*>` is used
to reference the node, which makes sure storage and access is managed in a way
that the same node can be accessed from multiple instances of the reference. It
also does user counting so the reference is only destroyed when entierly
unused, and keeps being updated until then.

It's crucial that widgets implement the `bwWidget::operator==()` function well
for this to work. It is what lets us recognize widgets over redraws.

With this, bWidgets shouldn't crash anymore on simple mouse movements, some
buttons will even do something on clicks. But state information like hover
feedback or scrollbar values is lost. Keeping that is the next step to get the
reconstruct-on-redraw logic to work.