Refactor: Change how widget handler functions access the widget
The code of each handler can now inherit from `WidgetEventHandler<>` and pass
it the widget type as template parameter, and then access the widget of this
exact type as `Widget()` member function. So the handlers don't have to
implement that getter themselves.
Some additional changes were needed for the in-between class
`bwAbstractButtonHandler`.

Also change make the `createHandler()` member function `const`.