Lazy create state object
Create state object on first request, so we don't rely on code using
`WidgetNew()` as opposed to allocating a widget in place (e.g. using `new` or
`std::make_unique()`). Technically `WidgetNew()` isn't needed anymore, but I'm
leaving it since it's a bit nicer than `std::make_unique()` and in future we
may need it again.

Also changed `bwWidget::createState()` to return the state object, rather than
setting it directly. This is safer since there's a single place where it's done
and can't be missed that way. It's also more consistent to `createHandler()`.