Initial new data-binding design
Wasn't quite happy with the apply functor design to change values in response
to widget interactions. I'm quite happy with this new design. Although the
implementation of the data-binding is a bit complicated (lots of template
shenanigans), using it is quite simple.

This is added on top of the `reconstruct-on-redraw` branch, for which this is
a followup basically.

Also added documentation, which will end up here:
https://julianeisel.github.io/bWidgets/md_docs_bWidgets_data_bindings.html